Windows ready Beta version

See merge request skander.hatira/bisepsgui!1

+++ src/main.js
@@ -1,15 +1,20 @@
 import { v4 as uuidv4 } from "uuid";
-const { app, BrowserWindow, Menu } = require("electron");
-const conda =
-  process.platform == "win32"
-    ? "conda"
-    : "$(head -n 1 $HOME/.conda/environments.txt)/bin/conda";
+const {
+  app,
+  BrowserWindow,
+  ipcMain,
+  session,
+  ipcRenderer,
+} = require("electron");
 const uid = uuidv4();
-const sock = `/tmp/bissprop${uid}.sock`;
 const mongod = require("./backend/spawnMongod.js");
 const os = require("os");
 const path = require("path");
 const fs = require("fs");
+console.log(os.tmpdir())
+const sock = process.platform == "win32" ?  path.join('\\\\?\\pipe', `biseps${uid}`)  :`/tmp/biseps${uid}.sock`
+const unixSocket =  process.platform == "win32" ? "mongodb://localhost:27017"  :`/tmp/bisepsMongod`
+
 const running = require("is-running");
 import installExtension, {
   REACT_DEVELOPER_TOOLS,
@@ -24,10 +29,12 @@ const mongodLock = path.join(
 const homedir = require("os").homedir();
 const bisepsTemp = path.join(homedir, ".bisepsTemp/");
 console.log(bisepsTemp);
+console.log(unixSocket)
+console.log(sock)
 console.log(process.platform);
 process.platform == "darwin" || process.platform == "linux"
   ? exec(
-      `${process.platform == "win32" ? "where conda" : "command -v conda"}`,
+      `${process.platform == "win32" ? "(Get-command conda).path" : "command -v conda"}`,
       (error, stdout, stderr) => {
         if (error) {
           console.log(`error: ${error.message}`);
@@ -62,13 +69,13 @@ process.platform == "darwin" || process.platform == "linux"
     });
 
 execSync(
-  `${conda} env create -f ${__dirname}/resources/${
+  `conda env create -f ${path.join(__dirname,"resources",
     process.platform == "darwin"
       ? "mongodbMac.yaml"
       : process.platform == "win32"
-      ? "mongodbWindows.yaml"
+      ? "mongoWindows.yaml"
       : "mongodbLinux.yaml"
-  } -n bisepsMongo || true`,
+  )} -n bisepsMongo ${process.platform == "win32" ?  "; $? -or $true" : ' || true ' } `,{"shell": process.platform == "win32" ? "powershell.exe" : "/bin/sh"},
   (error, stdout, stderr) => {
     if (error) {
       console.log(`error: ${error.message}`);
@@ -82,13 +89,13 @@ execSync(
   }
 );
 execSync(
-  `${conda} env create -f ${__dirname}/resources/${
-    process.platform == "darwin"
-      ? "snakemakeMac.yaml"
-      : process.platform == "win32"
-      ? "snakemakeWindows.yaml"
-      : "snakemakeLinux.yaml"
-  } -n bisepsSnakemake || true`,
+  `conda env create -f ${path.join(__dirname,"resources",
+  process.platform == "darwin"
+    ? "snakemakeMac.yaml"
+    : process.platform == "win32"
+    ? "snakemakeWindows.yaml"
+    : "snakemakeLinux.yaml"
+)} -n bisepsSnakemake ${process.platform == "win32" ?  '; $? -or $true' : ' || true ' }`,{"shell": process.platform == "win32" ? "powershell.exe" : "/bin/sh"},
   { shell: true, stdio: "inherit" },
   (error, stdout, stderr) => {
     if (error) {
@@ -105,7 +112,7 @@ execSync(
 if (fs.existsSync(mongodLock)) {
   fs.stat(mongodLock, function (err, stats) {
     if (stats.size === 0) {
-      mongod();
+      mongod(unixSocket);
     } else {
       fs.readFile(mongodLock, "utf8", function (err, data) {
         if (err) {
@@ -113,23 +120,24 @@ if (fs.existsSync(mongodLock)) {
         }
         if (running(data)) {
           console.log(
-            "database already running on /tmp/bisepsmongodb.sock pid : " + data
+            "database already running on bisepsmongodb.sock pid : " + data
           );
         } else {
           fs.unlinkSync(mongodLock);
-          mongod();
+          mongod(unixSocket);
         }
       });
     }
   });
 } else {
-  mongod();
+  mongod(unixSocket);
 }
+
 
 const server = require("../src/backend/spawnServer.js");
 
 // setTimeout(function () {
-server(sock);
+server(sock,unixSocket);
 // }, 4000);
 // const sock = "/tmp/bissprop.sock";
 
@@ -137,12 +145,53 @@ try {
   require("electron-reloader")(module);
 } catch (_) {}
 
+async function createJB(port) {
+  // Create the browser window.
+  // const newWindow = new BrowserWindow({
+  //   width: 1080,
+  //   height: 720,
+  //   webPreferences: {
+  //     nodeIntegration: true,
+  //     enableRemoteModule: true,
+  //     webSecurity: false,
+  //   },
+  // });
+  // Load app
+  // newWindow.webContents.loadURL(SECOND_WINDOW_WEBPACK_ENTRY);
+  // rest of code..
+}
 // Handle creating/removing shortcuts on Windows when installing/uninstalling.
 if (require("electron-squirrel-startup")) {
   // eslint-disable-line global-require
   app.quit();
 }
+// ipcMain.on("ping", (event, port) => {
+//   // console.log(port);
+//   // createJB(port);
+//   const newWindow = new BrowserWindow({
+//     width: 1080,
+//     height: 720,
+//     webPreferences: {
+//       preload: __dirname + "/preloadJB.js",
+//       nodeIntegration: false,
+//       nativeWindowOpen: true,
+//       nodeIntegrationInSubFrames: true,
+//       webSecurity: false,
+//     },
+//   });
+//   const dirname = "/home/Bureau/jbrowse2";
+//   // const url = require("url").format({
+//   //   protocol: "file",
+//   //   slashes: true,
+//   //   pathname: path.join(dirname, "worker.html"),
+//   // });
+//   // newWindow.loadURL(url);
 
+//   newWindow.loadURL(`http:///localhost:${port}`);
+//   newWindow.once("ready-to-show", () => {
+//     newWindow.show();
+//   });
+// });
 const createWindow = () => {
   // Create the browser window.
   const mainWindow = new BrowserWindow({
@@ -185,7 +234,9 @@ app.on("window-all-closed", () => {
     fs.rmdirSync(bisepsTemp, {
       recursive: true,
     });
-    fs.unlinkSync(sock);
+    if (process.platform !=="win32") {
+      fs.unlinkSync(sock);
+    }
     console.log("App Successfully Terminated");
     app.quit();
   }
